CwmsDataManager was setting office_id on its own.  Replaced with individual dao objects for consistency with all the other controller/dao pairs.
