+++ cwms_radar_api/src/main/java/cwms/radar/api/UnitsController.java
@@ -1,23 +1,24 @@
 package cwms.radar.api;
 
 import static com.codahale.metrics.MetricRegistry.name;
+import static cwms.radar.data.dao.JooqDao.getDslContext;
 
 import com.codahale.metrics.Histogram;
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.Timer;
 import cwms.radar.api.errors.RadarError;
-import cwms.radar.data.CwmsDataManager;
+import cwms.radar.data.dao.UnitsDao;
 import cwms.radar.formatters.Formats;
 import io.javalin.apibuilder.CrudHandler;
 import io.javalin.http.Context;
 import io.javalin.plugin.openapi.annotations.OpenApi;
 import io.javalin.plugin.openapi.annotations.OpenApiParam;
 import io.javalin.plugin.openapi.annotations.OpenApiResponse;
-import java.sql.SQLException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.servlet.http.HttpServletResponse;
+import org.jooq.DSLContext;
 
 public class UnitsController implements CrudHandler {
     private static final Logger logger = Logger.getLogger(UnitsController.class.getName());
@@ -69,8 +70,9 @@ public class UnitsController implements CrudHandler {
         getAllRequests.mark();
         try (
                 final Timer.Context timeContext = getAllRequestsTime.time();
-                CwmsDataManager cdm = new CwmsDataManager(ctx)
+                DSLContext dsl = getDslContext(ctx)
         ) {
+            UnitsDao dao = new UnitsDao(dsl);
             String format = ctx.queryParamAsClass("format", String.class).getOrDefault("json");
 
 
@@ -100,11 +102,11 @@ public class UnitsController implements CrudHandler {
                             + "implemented for this end point");
             }
 
-            String results = cdm.getUnits(format);
+            String results = dao.getUnits(format);
             ctx.status(HttpServletResponse.SC_OK);
             ctx.result(results);
             requestResultSize.update(results.length());
-        } catch (SQLException ex) {
+        } catch (Exception ex) {
             logger.log(Level.SEVERE, null, ex);
             ctx.status(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
             ctx.result("Failed to process request");
