--- cwms_radar_api/src/main/java/cwms/radar/data/CwmsDataManager.java
@@ -1,62 +1,0 @@
-package cwms.radar.data;
-
-import cwms.radar.ApiServlet;
-import java.sql.Connection;
-import java.sql.SQLException;
-import java.util.logging.Logger;
-
-import io.javalin.http.Context;
-import org.jooq.DSLContext;
-import org.jooq.SQLDialect;
-import org.jooq.impl.DSL;
-
-import usace.cwms.db.jooq.codegen.packages.CWMS_CAT_PACKAGE;
-import usace.cwms.db.jooq.codegen.packages.CWMS_ENV_PACKAGE;
-
-
-public class CwmsDataManager implements AutoCloseable {
-    private static final Logger logger = Logger.getLogger("CwmsDataManager");
-    public static final String FAILED = "Failed to process database request";
-
-    private Connection conn;
-    private DSLContext dsl;
-
-    public CwmsDataManager(Context ctx) throws SQLException{
-        this(ctx.attribute(ApiServlet.DATABASE), ctx.attribute(ApiServlet.OFFICE_ID));
-    }
-
-    public CwmsDataManager(Connection conn, String officeId) {
-        this.conn = conn;
-        dsl = DSL.using(conn, SQLDialect.ORACLE11G);
-
-        setOfficeId(officeId);
-    }
-
-    private void setOfficeId(String officeId)
-    {
-        CWMS_ENV_PACKAGE.call_SET_SESSION_OFFICE_ID(dsl.configuration(), officeId);
-    }
-
-    @Override
-    public void close() throws SQLException {
-        conn.close();
-    }
-
-
-	public String getUnits(String format) {
-        return CWMS_CAT_PACKAGE.call_RETRIEVE_UNITS_F(dsl.configuration(), format);
-	}
-
-	public String getParameters(String format){
-        return CWMS_CAT_PACKAGE.call_RETRIEVE_PARAMETERS_F(dsl.configuration(), format);
-    }
-
-	public String getTimeZones(String format) {
-            return CWMS_CAT_PACKAGE.call_RETRIEVE_TIME_ZONES_F(dsl.configuration(), format);
-    }
-
-
-
-
-
-}
